initial

+++ src/main/java/com/pnm/prices/infrastructure/persistence/repositories/BrandsRepository.java
@@ -3,6 +3,6 @@ package com.pnm.prices.infrastructure.persistence.repositories;
 import com.pnm.prices.infrastructure.persistence.entities.BrandEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface BrandsRepository extends CrudRepository<BrandEntity, Long> {
+public interface BrandsRepository extends CrudRepository<BrandEntity, Integer> {
 
 }
